refactor(geo): Move Cyl Face enum to Vol Bounds + python bindings
